+++ android/movies/app/src/main/res/layout/content_detail.xml
@@ -1,5 +1,5 @@
 <fragment android:id="@+id/fragment"
-		  android:name="net.usrlib.android.movies.DetailActivityFragment"
+		  android:name="net.usrlib.android.movies.fragment.DetailActivityFragment"
 		  xmlns:android="http://schemas.android.com/apk/res/android"
 		  xmlns:app="http://schemas.android.com/apk/res-auto"
 		  xmlns:tools="http://schemas.android.com/tools"
