+++ android/movies/app/src/main/res/layout/fragment_detail.xml
@@ -7,7 +7,7 @@
 				android:paddingLeft="@dimen/activity_horizontal_margin"
 				android:paddingRight="@dimen/activity_horizontal_margin"
 				android:paddingTop="@dimen/activity_vertical_margin"
-				tools:context="net.usrlib.android.movies.DetailActivityFragment"
+				tools:context=".fragment.DetailActivityFragment"
 				tools:showIn="@layout/activity_detail">
 
 	<ScrollView
@@ -55,13 +55,6 @@
 						android:layout_width="wrap_content"
 						android:layout_height="wrap_content">
 
-						<!--<TextView-->
-							<!--android:layout_width="wrap_content"-->
-							<!--android:layout_height="wrap_content"-->
-							<!--android:text="@string/movie_rating_label"-->
-							<!--android:textStyle="italic"-->
-							<!--android:layout_marginRight="@dimen/default_margin"-->
-							<!--style="@style/Base.TextAppearance.AppCompat.Medium"/>-->
 						<TextView
 							android:id="@+id/movie_rating"
 							android:layout_width="wrap_content"
@@ -74,25 +67,6 @@
 							android:text="@string/movie_rating_range"
 							style="@style/Base.TextAppearance.AppCompat.Small"/>
 					</LinearLayout>
-
-					<!--<LinearLayout-->
-						<!--android:orientation="horizontal"-->
-						<!--android:layout_width="wrap_content"-->
-						<!--android:layout_height="wrap_content">-->
-
-						<!--<TextView-->
-							<!--android:layout_width="wrap_content"-->
-							<!--android:layout_height="wrap_content"-->
-							<!--android:text="@string/movie_votes_label"-->
-							<!--android:textStyle="italic"-->
-							<!--android:layout_marginRight="@dimen/default_margin"-->
-							<!--style="@style/Base.TextAppearance.AppCompat.Medium"/>-->
-						<!--<TextView-->
-							<!--android:id="@+id/movie_votes"-->
-							<!--android:layout_width="wrap_content"-->
-							<!--android:layout_height="wrap_content"-->
-							<!--style="@style/Base.TextAppearance.AppCompat.Medium"/>-->
-					<!--</LinearLayout>-->
 
 					<LinearLayout
 						android:orientation="horizontal"
