+++ android/movies/app/src/main/res/layout/fragment_main.xml
@@ -6,7 +6,7 @@
 				android:paddingLeft="@dimen/activity_horizontal_margin"
 				android:paddingRight="@dimen/activity_horizontal_margin"
 				android:paddingTop="@dimen/activity_vertical_margin"
-				tools:context="net.usrlib.android.movies.MainActivityFragment"
+				tools:context=".fragment.MainActivityFragment"
 				android:background="@color/black"
 				tools:showIn="@layout/activity_main">
